+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Core: Implemented DMA functions
 - CRT: Separate log-related APIs from libc.a to liblog.a **(Disruptive)**
 - CRT: Added `vgs_memcpy`, `vgs_memset` and `vgs_strlen` _(Using DMA)_
+- CRT: Added `vgs_strchr` and `vgs_strrchr` _(Not DMA)_
 
 ## Version 0.2.0